requirements.txt
aiogram==3.*
python-dotenv
requests